requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.350
langchain-groq==0.0.1
langchain-community==0.0.1
faiss-cpu==1.7.4
chromadb==0.4.18
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
sentence-transformers==2.2.2
numpy==1.24.3 